Quote the arg given to shell command, in case it has spaces, etc
This argument is a path where tags should go (ie g:gutentags_cache_dir,
or the directory to scan if not set) and would need to be quoted if it
had spaces or other characters meaningful to the shell.  The variable
gets assessed by 'eval' before being used to invoke gtags, so we want to
embed the escapes within the string.

While we do this for correctness, note that unix cscope cannot handle
directories with spaces in the path to the tags file, because it puts
the directory name in the database and uses spaces to delimit records in
the first line (the filename, and also dash-prefixed options cscope had
when constructing the database, so there's no way to know for sure where
the filename ends).

Vim cannot handle spaces either.  The limitation does not exist for
gtags, but vim's ":cs add" (src/if_cscope.c::cs_add()), typically
invoked from gutentags_plus, does not handle spaces correctly itself.
It uses spaces in the arguments to break apart into significant
subwords, to distinguish other arguments to ':cs'.

Lastly the cscope command used by vim in src/if_cscope.c
cs_create_connection() does not properly quote the filename when it
passes to /bin/sh -c "exec cscope -dl -f %s" so that wouldn't handle
spaces either.

For now we can live with the limitation that we can't ever use cscope
functionality in paths with spaces in them.