Set noclobber in update scripts, otherwise lock file is useless
The shell default is for clobber to be set, so "echo $$ > $LOCKFILE"
will succeed no matter how many times it's run.  In other words, this
wasn't acting as a lockfile at all, unless the user had set 'noclobber'
in their shell's non-interactive profile.

In each of these scripts, a trap exists that does cleanup, so the lock
files should get removed at script exit no matter what.  If the computer
crashed or lost power in between though, the stale lock file would be
left and have to be manually removed.  This is not unlike any other
stale lock file issue.

However we could check if the pid exists, since it's written to the
file.  That sounds invasive and non-portable... maybe later.