Remove even partial files as they can cause gtags segfault
If we exit early before finish, sometimes partial files are left around.
Unfortunately these can cause gtags to segfault.  "incremental" in gtags
should not be thought of as "start where we left off when the last job
was prematurely killed" but rather "incrementally update the [previously
completely-generated] database by scanning only files that have
changed."  If the db was not fully generated, we must remove all the
partial files from within the trap.  This might cause a lot of work to
be redone in large source bases, but gtags demonstrably segfaults on
partials in some cases, and refuses to continue in other cases, claiming
that the db is corrupt.  It needs to finish, or its results need to be
discarded, but not anything in-between.

All that is to say, we will not use 'test -s' and leave a partial
around, but just simply remove any files if the handler is entered at
all.

We must remove the EXIT from the trap-signals list then, because if we
leave it, it will always be removed even if fully complete!